[Added] ABC-008~027

+++ AtCoder/Libraries/数字系/各桁の和.cpp
@@ -3,7 +3,7 @@ using namespace std;
 
 typedef long long ll;
 
-int GetSumOfDigits(ll N) {
+ll GetSumOfDigits(ll N) {
     ll sum = 0;
     while (N > 0) {
         sum += N % 10;
